feat: Add save functions for moves and items
This commit adds save functions for moves and items in the AttackSaver class. The `SaveMove` function saves the move used by a character, while the `SaveItem` function saves the item used by a character. These changes improve the functionality of the game by allowing the tracking of character actions in battle.