```text feat: Add currentSeed and currentGroupSet fields to BattleSystem
This commit adds the currentSeed and currentGroupSet fields to the BattleSystem class. These fields are used to keep track of the current seed and group set for enemy setup in the battle. This change enhances the functionality of the battle system by allowing the instantiation of enemy characters based on the current seed and group set.