Refactor ToolObject class to add debuffs property
This commit adds a new serialized field, debuffs, and debuffAmount to the ToolObject class in the Assets/Scripts/Inventory Objects/ToolObject.cs file. The debuffs property is used to store the types of debuffs that can be applied to the tool, and the debuffAmount property represents the number of debuffs that can be applied. This enhancement improves the functionality and customization options for tools in the game.