Refactor Moves class to add buff and debuff values
This commit adds two new serialized fields, buffValue and debuffValue, to the Moves class in the Assets/Scripts/Moves.cs file. These fields are used to store the values of buffs and debuffs that can be applied to moves. This enhancement improves the functionality and customization options for moves in the game.