Refactor CharacterTemplate class to add debuff removal functionality
This commit adds a new method, RemoveDebuffs, to the CharacterTemplate class in the Assets/Scripts/CharacterTemplate.cs file. The method takes an array of HealingMoves.HealDebuff enum values as a parameter and removes the corresponding debuffs from the character's stats. This enhancement improves the flexibility and functionality of the debuff system in the game.