Refactor Movesets class to update menu name for new moveset
This commit updates the Movesets class in the Assets/Scripts/Moves/Movesets.cs file to change the menu name for creating a new moveset. The previous menu name was "Moves/New Moveset" and it has been updated to "Moves/Moveset/New Moveset". This change improves the organization and clarity of the menu structure for movesets.