Refactor CharacterTemplate class to fix character status bug
This commit fixes a bug in the CharacterTemplate class where the characterStatus was not being updated correctly when the currentHP reached 0. The bug was caused by a commented out line of code that set the characterStatus to CharacterStatus.Downed. This line of code has been uncommented to correctly update the characterStatus. This fix ensures that the character status is accurately reflected when their HP reaches 0.