Refactor BattleSystem.cs to instantiate enemy health bars
This commit modifies the BattleSystem.cs file to instantiate enemy health bars. It adds a new method, InstantiateEnemyHealthBars, to the UIManager class, which takes a list of enemy members and creates health bars for each enemy. This change improves the user interface of the battle system by providing visual representation of enemy health during battles.