```text feat: Add OnLevelUp event to EventController.cs
This commit adds the OnLevelUp event to the EventController.cs file. The OnLevelUp event will be triggered when a character levels up in the game. This change enhances the functionality of the EventController class by providing a way to handle level up events and perform additional actions, such as playing animations or sound effects.